requirements.txt
streamlit
telethon
crewai
langchain
python-dotenv
sqlalchemy
chromadb
openai
nest_asyncio
pydantic 